requirements.txt
fastapi==0.115.0
uvicorn==0.31.1
python-dotenv==1.0.1
firebase-admin==6.5.0
pyodbc==5.1.0
azure-storage-queue==12.11.0